app/api/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
requests ==2.31.0